Rename list-networks.* to list_networks.*
