+++ src/client/cli/client.cpp
@@ -23,8 +23,8 @@
 #include "cmd/help.h"
 #include "cmd/info.h"
 #include "cmd/launch.h"
-#include "cmd/list-networks.h"
 #include "cmd/list.h"
+#include "cmd/list_networks.h"
 #include "cmd/mount.h"
 #include "cmd/purge.h"
 #include "cmd/recover.h"
