+++ src/client/cli/cmd/list_networks.h
